Overhaul

+++ driver.py
@@ -6,14 +6,28 @@ import sys
 import choose as c
 import utilities as utils
 
+#
+# Stuff to use in runtime
+#
+prefix = 'c!'
+response_timeout = 5
+confirm_emoji = "\u2705"
+deny_emoji = "\u26d4"
+commands = {
+    'help': 'Stop it, get some help.',
+    'choose <[Int] players> <[Int] civilizations>': 'Gets a random set of civilizations from n number of players. Players field is mandatory, civilizations field is optional.'
+}
+DEFAULT_BOT_STATUS = 'Twiddling thumbs...'
+website = 'https://github.com/dmitri-mcguckin/civ-randomizer'
+logo_url = 'https://i.imgur.com/pBgN2gZ.jpg'
+
 class ImposibleReactionResolve(Exception):
     def __init__(self):
         utils.log(3, "Reaction confirmation could not resolve!")
 
-def bot_status():
-    global DEFAULT_BOT_STATUS
+def bot_status(message=DEFAULT_BOT_STATUS):
     global prefix
-    return (prefix + "help" + " | " + DEFAULT_BOT_STATUS)
+    return (prefix + "help" + " | " + message)
 
 async def handle_forced_exit():
     global client
@@ -35,6 +49,20 @@ def message_argument(message):
 def is_command(message):
     return (message_prefix(message) == prefix)
 
+class Timer:
+    def __init__(self, timeout, callback):
+        self._timeout = timeout
+        self._callback = callback
+        self._task = asyncio.ensure_future(self._job())
+
+    async def _job(self):
+        await asyncio.sleep(self._timeout)
+        utils.log(1, 'TIMEOUT EVENT!')
+        await self._callback()
+
+    def cancel(self):
+        self._task.cancel()
+
 async def set_status(status_message, new_status):
     global client
     game = discord.Game(status_message)
@@ -47,10 +75,9 @@ async def resolve_confirmation(message, question):
     global deny_emoji
 
     def reactor_is_author(reaction, user):
-        global debug
         global confirm_emoji
         global deny_emoji
-        if(debug): utils.log(2, 'Confirm: ' + str(str(reaction) == confirm_emoji) + '\t\tDeny: ' + str(str(reaction) == deny_emoji))
+        if(utils.DEBUG): utils.log(2, 'Confirm: ' + str(str(reaction) == confirm_emoji) + '\t\tDeny: ' + str(str(reaction) == deny_emoji))
         return (user == message.author and ((str(reaction.emoji) == confirm_emoji) or (str(reaction.emoji) == deny_emoji)))
 
     response = await message.channel.send(content=question)
@@ -83,79 +110,50 @@ async def usage(message):
     for command in commands.keys():
         help_message.add_field(name=str(prefix + command), value=str(commands[command]), inline=False)
 
-    if(debug):
+    if(utils.DEBUG):
         utils.log(2, '' + str(message.channel) + '\t' + str(message.mention_everyone))
 
     await message.channel.send(content=None, embed=help_message)
 
 async def choose(message):
     global client
+    global randomizer
 
     argv = message.content.split(' ')
     argc  = len(argv)
-    player_civs = {}
+    results = {}
 
-    try:
-        if(argc == 2): player_civs = c.pChoose(int(argv[1]))
-        elif(argc == 3): player_civs = c.pChoosec(int(argv[1]), int(argv[2]))
-    except c.GreedyPlayer:
-        try:
-            result = await resolve_confirmation(message, 'No civilization pool size was specified, but recommended size is: ' + str(c.recommended_pool_size(argv[1])) + '. Would you like to use this setting?')
-
-            if(result):
-                c.reset_pool()
-                utils.log(2, 'Players: ' + argv[1] + '\tCivilizations: ' + str(c.recommended_pool_size(argv[1])))
-                player_civs = c.pChoosec(int(argv[1]), c.recommended_pool_size(argv[1]))
-            else:
-                await message.channel.send(content='**Canceling dice roll!**')
-                message.delete()
-                return
-        except ImposibleReactionResolve as e:
-            utils.log(3, 'Could not resolve reaction: ' + str(e) + '\n\tThrowing out thread!')
-            await message.channel.send(content='**Canceling dice roll!**')
-            await message.delete()
-            return
+    if(argc == 2): results = randomizer.choose(int(argv[1]))
+    elif(argc == 3): results = randomizer.choose(int(argv[1]), civilization_count=int(argv[2]))
 
     response = discord.Embed(title='Civilization Randomizer Results', url=website, description='Results of the randomizer.', color=0x58ff00)
     response.set_author(name='Civilization Randomizer', url=website, icon_url=logo_url)
     response.set_thumbnail(url=logo_url)
 
     utils.log(0, 'Player Civilization selections:')
-    for player in range(0,len(player_civs)):
-        utils.log(0, '\tPlayer ' + str(player) + ': ' + str(player_civs[player]))
-        response.add_field(name=str('Player ' + str(player + 1) + ':'), value=str('Choose from: ' + str(player_civs[player])), inline=False)
+    for player,civilizations in results.items():
+        civilization_string = "Choose from: "
+        for i in civilizations:
+            if(i != civilizations[-1]): civilization_string += i + ', '
+            else: civilization_string += 'or ' + i
+        response.add_field(name=player, value=civilization_string, inline=False)
 
-    await message.delete()
+    if(utils.DELETE_AFTER_PROCESS): await message.delete()
     await message.channel.send(content=None, embed=response)
 
 async def bad_command(message):
     global client
 
     utils.log(1, 'Cannot process this command: ' + message.content + '\n\tIgnoring!')
-    await message.delete()
+    if(utils.DELETE_AFTER_PROCESS): await message.delete()
     await message.channel.send(content=('**<@' + str(message.author.id) + '> Bad command: \"' + message_prefix(message.content) + message_suffix(message.content) + '\"!** *(Try using ' + message_prefix(message.content) + "help for correct usage)*"))
 
 #
 # Initialization
 #
 client = discord.Client()
+randomizer = c.CivRandomizer()
 bot_token = os.getenv('CIV_BOT_TOKEN')
-
-#
-# Stuff to use in runtime
-#
-debug = True
-prefix = 'c!'
-response_timeout = 5
-confirm_emoji = "\u2705"
-deny_emoji = "\u26d4"
-commands = {
-    'help': 'Stop it, get some help.',
-    'choose <[Int] players> <[Int] civilizations>': 'Gets a random set of civilizations from n number of players. Players field is mandatory, civilizations field is optional.'
-}
-DEFAULT_BOT_STATUS = 'Twiddling thumbs...'
-website = 'https://github.com/dmitri-mcguckin/civ-randomizer'
-logo_url = 'https://i.imgur.com/pBgN2gZ.jpg'
 
 #
 # Discord event stuff
@@ -163,11 +161,10 @@ logo_url = 'https://i.imgur.com/pBgN2gZ.jpg'
 @client.event
 async def on_ready():
     utils.log(0, 'Client logged in as the user ' + str(client.user) + '\n\tID: ' + str(client.user.id) + '\n\tBOT: ' + str(client.user.bot) + '\n\tDisplay Name: ' + str(client.user.display_name))
-    await set_status(DEFAULT_BOT_STATUS, discord.Status.idle)
+    await set_status(bot_status(DEFAULT_BOT_STATUS), discord.Status.online)
 
 @client.event
 async def on_message(message):
-    c.reset_pool()
     # Check if message has correct prefix
     if(is_command(message.content)):
         # Console confirmation that a relevant command was received
@@ -177,13 +174,13 @@ async def on_message(message):
         command = message_suffix(message.content)
 
         if(command == "help"):
-            await set_status('Getting help...', discord.Status.dnd)
+            await set_status(bot_status('Getting help...'), discord.Status.dnd)
             await usage(message)
         elif(command == 'choose'):
-            await set_status('Rolling dice...', discord.Status.dnd)
+            await set_status(bot_status('Rolling dice...'), discord.Status.dnd)
             await choose(message)
         else:
-            await set_status('Invalid command!', discord.Status.dnd)
+            await set_status(bot_status('Invalid command!'), discord.Status.dnd)
             await bad_command(message)
 
         await set_status(bot_status(), discord.Status.online)
